+++ packages/nana-ui/src/shell/index.ts
@@ -1,3 +1,4 @@
 export { default as NanaAppShell } from "./NanaAppShell.vue";
+export { default as NanaDesktopShell } from "./NanaDesktopShell.vue";
 export { default as NanaSidebar } from "./NanaSidebar.vue";
 export { default as NanaTitleBar } from "./NanaTitleBar.vue";