+++ packages/ui/src/shell.ts
@@ -1,6 +1,7 @@
 /// <reference path="./shims-vue.d.ts" />
 
 export { default as LiliaAppShell } from "./layouts/AppShell.vue";
+export { default as LiliaDesktopShell } from "./layouts/LiliaDesktopShell.vue";
 export { default as LiliaSidebarFrame } from "./layouts/LiliaSidebarFrame.vue";
 export { default as LiliaSidebarNavRow } from "./components/sidebar/SidebarNavRow.vue";
 export { default as LiliaSidebarFooter } from "./components/sidebar/SidebarFooter.vue";